backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.2
PyYAML==6.0.2
sqlalchemy==2.0.34
aiosqlite==0.20.0
python-multipart==0.0.9
greenlet==3.1.1
